+++ projects/ng-gallery/src/lib/components/gallery-dots.component.ts
@@ -1,4 +1,5 @@
 import { Component, Input, ChangeDetectionStrategy } from '@angular/core';
+import { NgFor } from '@angular/common';
 import { Gallery } from '../services/gallery.service';
 import { GalleryState } from '../models/gallery.model';
 import { GalleryConfig } from '../models/config.model';
@@ -15,7 +16,9 @@ import { GalleryConfig } from '../models/config.model';
          (click)="gallery.ref(this.galleryId).set(i)">
       <div class="g-dot-inner"></div>
     </div>
-  `
+  `,
+  standalone: true,
+  imports: [NgFor]
 })
 export class GalleryDotsComponent {
   @Input() galleryId: string;